functions/aid_ingestion/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.47.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
googlemaps>=4.10.0
python-json-logger>=2.0.7
